extracted config.Validate into a separate package
